% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointCureModel.R
\name{jointCureModel}
\alias{jointCureModel}
\title{Estimation of the joint longitudinal time-to-event models including a cure fraction}
\usage{
jointCureModel(formFixed, formRandom, formLatency, formIncidence, formID,
  IdVar, data, timeVar, jointCureModel = c("FJCmodel"),
  survMod = c("weibull-PH"), param = c("shared-RE"),
  Infprior_cure = F, smcure_out = NULL, classif_trick = TRUE,
  cov_prior = c("inverse-gamma"), Sigma_d = FALSE, n.chains = 1,
  n.iter = 10000, n.burnin = 5000, n.thin = 1, n.adapt = 5000,
  C = 1, priorTau = 100, out_data = T)
}
\arguments{
\item{formFixed}{formula for fixed part of longitudinal submodel with response variable}

\item{formRandom}{formula for random part of longitudinal submodel without response variable}

\item{formLatency}{survival formula as formula in survival package for latency submodel}

\item{formIncidence}{formula specifying covariate in incidence submodel}

\item{formID}{formula specifying the ID variable (e.g. = ~ subject)}

\item{IdVar}{string specify the names of ID variable (Id subject)}

\item{data}{dataset of observed variables}

\item{timeVar}{string specify the names of time variable (time of repeated measurements)}

\item{jointCureModel}{joint cure model to estimate including: "MCmodel","JLCCmodel","FJCmodel"}

\item{survMod}{form of survival submodel (only "weibull-PH" is available until now)}

\item{param}{shared association for "FJCmodel": "shared-RE" (default) or "td-value"}

\item{Infprior_cure}{most of prior distributions are data-driven if 'TRUE', otherwise vague prior distribution are considered}

\item{smcure_out}{estimated model from smcure package, otherwise is NULL by default}

\item{classif_trick}{if TRUE, subjects are cured if t_obs>max(t_obs(delta==1), otherwise D is only know for subjects having developed the event Taylor's rule about the cure status, D==0 if T_obs>max(t_obs[which(delta==1)])}

\item{cov_prior}{="wishart" prior distribtion for random effect covariance matrix or ="inverse-gamma" for independent prior distribtions with random effect variance matrix}

\item{Sigma_d}{if TRUE, the random effect matrix is class-specific, otherwise is the matrix is defined for the whole population}

\item{n.chains}{the number of parallel chains for the model; default is 1.}

\item{n.iter}{nteger specifying the total number of iterations; default is 10000}

\item{n.burnin}{integer specifying how many of n.iter to discard as burn-in ; default is 5000}

\item{n.thin}{integer specifying the thinning of the chains; default is 1}

\item{n.adapt}{integer specifying the number of iterations to use for adaptation; default is 5000}

\item{C}{positive integer for the zero trick used to define the likelihood in JAGS, default is 1}

\item{priorTau}{variance by default for vague prior distribution}

\item{out_data}{Boolean such as TRUE if you want the data of different submodels in output or FALSE otherwise}
}
\value{
A 'JMcuR' object which is a list with the following elements:
   \describe{
  \item{\code{Coefficients}}{list of posterior mean of each parameter}
  \item{\code{Modes}}{ list of posterior modes compute from the posterior samples of the parameters}
  \item{\code{Sd}}{list of Standard deviation of each parameters}
  \item{\code{sims.list}}{list of the MCMC chains of the parameters}
   }
}
\description{
Function using JAGS to estimate the joint longitudinal survival models with cure fraction: mixutre cure model (MCmodel) [Farewell, 1982], joint latent class cure model (JLCCmodel),  full joint cure model (FJCmodel) [Law et al. 2002, YU et al. 2004,2008]
}
\examples{

## estimation of the MCM for parameter initialisation
aids.id <- unique(aids[,c("id","time","death","drug","gender","prevOI","AZT")])
table(aids.id$death)
aids.id2 <- aids.id
aids.id2$drug <- as.numeric(aids.id$drug)-1
aids.id2$gender <- as.numeric(aids.id$gender)-1
aids.id2$prevOI <- as.numeric(aids.id$prevOI)-1
aids.id2$AZT <- as.numeric(aids.id$AZT)-1
smcure_out <- smcure(Surv(time, death) ~ drug + gender + prevOI + AZT,
                     cureform=~ drug + gender + prevOI + AZT,
                     data = aids.id2,
                     model="ph")

## Estimation of the joint latent class cure model (JLCCM)
JLCCM <- jointCureModel(formFixed = CD4 ~ obstime + drug + gender + prevOI + AZT,
                        formRandom = ~ obstime,
                        timeVar= "obstime",
                        formLatency = Surv(time, death) ~ drug + gender + prevOI + AZT,
                        formIncidence = ~ drug + gender + prevOI + AZT,
                        formID= ~ id,
                        IdVar = "id",
                        data = aids,
                        # model specifications
                        jointCureModel = c("JLCCmodel"),
                        survMod = c("weibull-PH"),
                        param = c("shared-RE"),
                        # prior options
                        Infprior_cure = TRUE,
                        smcure_out = smcure_out,
                        priorTau = 100,
                        # classification options
                        classif_trick = TRUE,
                        cov_prior = c("inverse-gamma"),
                        Sigma_d = TRUE)

## details of the estimated model
summary(JLCCM)

## Estimation of the full joint cure model given shared curent value (FJCM_A1)
FJCM_A1 <- cureJMbayes(formFixed = CD4 ~ obstime,
                       formRandom = ~ obstime,
                       timeVar= "obstime",
                       formLatency = Surv(time, death) ~ drug + gender + prevOI + AZT,
                       formIncidence = ~ drug + gender + prevOI + AZT,
                       formID= ~ id,
                       IdVar = "id",
                       data = aids,
                       # model specifications
                       jointCureModel = c("FJCmodel"),
                       survMod = c("weibull-PH"),
                       param = c("td-value"),
                       # prior options
                       Infprior_cure = TRUE,
                       smcure_out = smcure_out,
                       priorTau = 100,
                       # classification options
                       classif_trick = TRUE,
                       cov_prior = c("inverse-gamma"),
                       Sigma_d = TRUE,
                       # MCMC options
                       n.chains = 1, n.iter = 10000, n.burnin = 5000, n.thin = 1, n.adapt = 5000, quiet = FALSE, C = 1,
                       # out option
                       out_data=T)

## details of the estimated model
summary(FJCM_A1)

## Estimation of the full joint cure model given shared curent value (FJCM_A1)
FJCM_A2 <- cureJMbayes(formFixed = CD4 ~ obstime,
                       formRandom = ~ obstime,
                       timeVar= "obstime",
                       formLatency = Surv(time, death) ~ drug + gender + prevOI + AZT,
                       formIncidence = ~ drug + gender + prevOI + AZT,
                       formID= ~ id,
                       IdVar = "id",
                       data = aids,
                       # model specifications
                       jointCureModel = c("FJCmodel"),
                       survMod = c("weibull-PH"),
                       param = c("shared-RE"),
                       # prior options
                       Infprior_cure = TRUE,
                       smcure_out = smcure_out,
                       priorTau = 100,
                       # classification options
                       classif_trick = TRUE,
                       cov_prior = c("inverse-gamma"),
                       Sigma_d = TRUE,
                       # MCMC options
                       n.chains = 1, n.iter = 10000, n.burnin = 5000, n.thin = 1, n.adapt = 5000, quiet = FALSE, C = 1,
                       # out option
                       out_data=T)

## details of the estimated model
summary(FJCM_A2)
}
