% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.JMcuR.R
\name{predict.JMcuR}
\alias{predict.JMcuR}
\title{Predictions of the cure membership, the longitudinal or survival responses}
\usage{
\method{predict}{JMcuR}(object, newdata = NULL, type = c("subject"),
  state = c("cure"), level.interval = 0.95, Tsurv = NULL,
  Yt = NULL, yVar = "y", idVar = "subject", scale = 1.6,
  MCMC = TRUE, M = NULL, ...)
}
\arguments{
\item{object}{an object inheriting from class 'JMcuR'}

\item{newdata}{a data frame in which to look for variables with which to predict (default is NULL, and the prediction is done on the data used to estimate the cure model)}

\item{type}{a character string indicating the type of predictions to compute, marginal or subject-specific.}

\item{state}{a (vector of) character string indicating the response ("cure","survival","longitudinal") to predict}

\item{level.interval}{a numeric scalar denoting the tolerance/confidence level for the credibility interval}

\item{Tsurv}{A numeric scalar denoting the time of interest associated with the prediction; 'Tsur' is NULL by default and then the censoring time is considered to predict the cure status.}

\item{Yt}{a numeric scalar denoting the value of the longitudinal measurement if type='profile'. yt is NULL by default}

\item{yVar}{a character string indicating the name of the variable in newdata that corresponds to the longitudinal outcome;}

\item{idVar}{a character string indicating the name of the variable in newdata that corresponds to the subject identifier; required when type = "Subject"}

\item{scale}{a numeric value setting the scaling of the covariance matrix of the empirical Bayes estimates in the Metropolis step during the Monte Carlo sampling.}

\item{MCMC}{logical; if TRUE prediction procedure is done using Bayesian approach, and if FALSE the Frequentist approach is then considered for the prediction procedure}

\item{M}{a numerical scalar denoting the sampling size using Bayesian approach; is NULL by default and the lengh of MC chains in estimation step is then considered.}

\item{...}{further arguments to be passed to or from other methods. They are ignored in this function.}
}
\value{
An object 'curePredict' being a list with the following elements:
   \describe{
  \item{\code{out_pred}}{Dataframe with the (subject-specific) prediction results.}
  \item{\code{sims.list}}{list of the MCMC chains of the parameters}
  \item{\code{data}}{the argument 'newdata' used}
  \item{\code{type}}{the argument 'type' used}
  \item{\code{state}}{the argument 'state' used}
   }
}
\description{
Predictions of the cure membership, the longitudinal or survival responses from the estimated joint longitudinal survival models with cure fraction
}
\examples{

}
