% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.JMcuR.R
\name{summary.JMcuR}
\alias{summary.JMcuR}
\title{Summary of a 'JMcuR' objects}
\usage{
\method{summary}{JMcuR}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class 'JMcuR'}

\item{...}{further arguments to be passed to or from other methods. They are ignored in this function.}
}
\value{
Returns NULL.
}
\description{
The function provides a summary of jointCureModel estimations.
}
