% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.JMcuR.R
\name{predict.JMcuR}
\alias{predict.JMcuR}
\title{Predictions of the cure membership, the longitudinal or survival responses}
\usage{
\method{predict}{JMcuR}(
  object,
  newdata = NULL,
  type = "subject",
  state = "cure",
  level.interval = 0.95,
  Tsurv = NULL,
  Yt = NULL,
  yVar = "y",
  idVar = "subject",
  MCMC = FALSE,
  M = NULL,
  ...
)
}
\arguments{
\item{object}{an object inheriting from class \code{JMcuR}}

\item{newdata}{a data frame in which to look for variables with which to predict (default is NULL, and the prediction is done on the data used to estimate the cure model)}

\item{type}{a character string indicating the type of predictions to compute, marginal or subject-specific.}

\item{state}{a (vector of) character string indicating the response ("cure","survival","longitudinal") to predict}

\item{level.interval}{a numeric scalar denoting the tolerance/confidence level for the credibility interval; the credible interval is return only for \cite{MCMC=TRUE}}

\item{Tsurv}{A numeric scalar denoting the time of interest associated with the prediction; 'Tsur' is NULL by default and then the censoring time is considered to predict the cure status.}

\item{Yt}{a numeric scalar denoting the value of the longitudinal measurement if type='profile'. yt is NULL by default}

\item{yVar}{a character string indicating the name of the variable in newdata that corresponds to the longitudinal outcome;}

\item{idVar}{a character string indicating the name of the variable in newdata that corresponds to the subject identifier; required when type = "Subject"}

\item{MCMC}{logical; if TRUE prediction procedure is done using Bayesian approach, and if FALSE (by default) the Frequentist approach is then considered for the prediction procedure}

\item{M}{a numerical scalar denoting the sampling size using Bayesian approach; is NULL by default and the lengh of MC chains in estimation step is then considered.}

\item{...}{further arguments to be passed to or from other methods. They are ignored in this function.}
}
\value{
An object \code{curePredict} being a list with the following elements:
   \describe{
  \item{\code{out_pred}}{Dataframe with the (subject-specific) prediction results.}
  \item{\code{sims.list}}{list of the MCMC chains of the parameters}
  \item{\code{data}}{the argument 'newdata' used}
  \item{\code{type}}{the argument 'type' used}
  \item{\code{state}}{the argument 'state' used}
   }
}
\description{
Predictions of the cure membership, the longitudinal or survival responses from the estimated joint longitudinal survival models with cure fraction
}
\examples{

## For the exemple(s), use the data 'aids' from joineR package
data("aids", package = "joineR")

## estimation of the MCM for parameter initialisation
aids.id <- unique(aids[,c("id","time","death","drug","gender","prevOI","AZT")])
aids.id2 <- aids.id
aids.id2$drug <- as.numeric(aids.id$drug)-1
aids.id2$gender <- as.numeric(aids.id$gender)-1
aids.id2$prevOI <- as.numeric(aids.id$prevOI)-1
aids.id2$AZT <- as.numeric(aids.id$AZT)-1
smcure_out <- smcure(Surv(time, death) ~ drug + gender + prevOI + AZT,
                     cureform=~ drug + gender + prevOI + AZT,
                     data = aids.id2,
                     model="ph")

## Estimation of the joint latent class cure model (JLCCM)
JLCCM <- jointCureModel(formFixed = CD4 ~ obstime + drug + gender + prevOI + AZT,
                        formRandom = ~ obstime,
                        timeVar= "obstime",
                        formLatency = Surv(time, death) ~ drug + gender + prevOI + AZT,
                        formIncidence = ~ drug + gender + prevOI + AZT,
                        formID= ~ id,
                        IdVar = "id",
                        data = aids,
                        # model specifications
                        jointCureModel = "JLCCmodel",
                        survMod = "weibull-PH",
                        param = "shared-RE",
                        # prior options
                        n.iter = 1000,
                        n.burnin = 500,
                        Infprior_cure = TRUE,
                        smcure_out = smcure_out,
                        priorTau = 100,
                        # classification options
                        classif_trick = TRUE,
                        cov_prior = "inverse-gamma",
                        Sigma_d = TRUE)

## details of the estimated model
summary(JLCCM)

## prediction step with the JLCCM
# Frequentist approach
pred_JLCCM <- predict(object = JLCCM,
                      newdata = NULL,
                      type = "subject",
                      state = c("cure"),
                      level.intervalle=0.95,
                      Tsurv = NULL,
                      Yt = NULL,
                      yVar = "CD4",
                      idVar = "id",
                      MCMC = FALSE,
                      M = NULL)

\dontrun{
# Bayesian approach
pred_JLCCM_MCMC <- predict(object = JLCCM,
                           newdata = NULL,
                           type = "subject",
                           state = c("cure"),
                           level.intervalle=0.95,
                           Tsurv=NULL,
                           Yt = NULL,
                           yVar = "CD4",
                           idVar = "id",
                           MCMC = TRUE,
                           M = 500)
}

}
\seealso{
\code{\link{jointCureModel}}
}
\author{
Antoine Barbieri and Catherine Legrand
}
