% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.JMcuR.R
\name{summary.JMcuR}
\alias{summary.JMcuR}
\title{Summary of a 'JMcuR' objects}
\usage{
\method{summary}{JMcuR}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class 'JMcuR'}

\item{...}{further arguments to be passed to or from other methods. They are ignored in this function.}
}
\value{
Returns NULL.
}
\description{
The function provides a summary of jointCureModel estimations. First, joint cure model's details are given.
Next, the estimation parameters including the standard erreur, standard deviation, and bound of the
credible interval at 95% are given for incidence submodel, latency submodel and conditionnal longitudinal submodels,
respectively.
}
\seealso{
\code{\link{jointCureModel}}
}
\author{
Antoine Barbieri, Catherine Legrand
}
